Fix: score consistency
