+++ src/components/molecules/faceRecognition.tsx
@@ -6,30 +6,44 @@ import {
   TNetInput,
 } from 'face-api.js';
 import useInterval from 'use-interval';
-import { MoviePlayerState } from 'types';
 import { Movie } from 'types/dataTypes';
 import { useMediaQuery } from '@mui/material';
 import { motion } from 'framer-motion';
 import { ScreenSize } from 'components/constants';
 import { CoreFunctionsContext } from 'contexts/coreFunctionsContext';
+import { axiosDefaultInstance } from 'utils/axiosConfig';
+import { useAuth0 } from '@auth0/auth0-react';
 
 type FaceRecognitionProps = {
   movie: Movie;
   movieDetailRef: any;
 };
 
+type PutMovieBody = {
+  grinningScore: string;
+};
 // recognition threshold for determining expression
-const THRESHOLD = 0.5;
+const THRESHOLD = 0.3;
+const RECOGNITION_INTERVAL = 250;
+const MAX_BATCH_ITER = 40;
 
-const FaceRecognition: FC<FaceRecognitionProps> = ({ movieDetailRef }) => {
+const FaceRecognition: FC<FaceRecognitionProps> = ({
+  movieDetailRef,
+  movie,
+}) => {
   const { grinningScore, setGrinningScore, moviePlayerState } =
     useContext(CoreFunctionsContext);
+
+  const { getAccessTokenSilently } = useAuth0();
 
   const webcamRef = useRef<HTMLVideoElement>(null);
   const [isModelReady, setIsModelReady] = useState<boolean>(false);
   const [isWebcamReady, setIsWebcamReady] = useState<boolean>(false);
 
   const [isDragging, setIsDragging] = useState<boolean>(false);
+
+  const [batchIter, setBatchIter] = useState<number>(0);
+  const [batchScore, setBatchScore] = useState<number>(0);
 
   const isLargerThanIpad = useMediaQuery(ScreenSize.largerThanIpad);
   const cameraSize = isLargerThanIpad ? '5rem' : '3rem';
@@ -54,11 +68,21 @@ const FaceRecognition: FC<FaceRecognitionProps> = ({ movieDetailRef }) => {
         )) {
           if (value > THRESHOLD && key == 'happy') {
             setGrinningScore(grinningScore + 1);
+            setBatchScore(batchScore + 1);
           }
         }
       }
+      // each batch includes MAX_BATCH_ITER number of recognition iteration
+      if (batchIter === MAX_BATCH_ITER && batchScore > 0) {
+        await handleSendScore(batchScore);
+        setBatchIter(0);
+        setBatchScore(0);
+        return;
+      }
+      // keep track of the number of iterations
+      setBatchIter(batchIter + 1);
     }
-  }, 300);
+  }, RECOGNITION_INTERVAL);
 
   const handleLoadModels = async () => {
     await faceapi.loadTinyFaceDetectorModel('/models');
@@ -74,8 +98,8 @@ const FaceRecognition: FC<FaceRecognitionProps> = ({ movieDetailRef }) => {
         video: { width: 100, height: 100 },
       });
 
-      webcam.srcObject = mediaStream;
-      setIsWebcamReady(true);
+      webcam.srcObject = await mediaStream;
+      await setIsWebcamReady(true);
     }
   };
 
@@ -93,6 +117,23 @@ const FaceRecognition: FC<FaceRecognitionProps> = ({ movieDetailRef }) => {
     }
   };
 
+  const handleSendScore = async (score: number) => {
+    const putBody: PutMovieBody = { grinningScore: `${score}` };
+    try {
+      const token = await getAccessTokenSilently({
+        audience: process.env.NEXT_PUBLIC_AUTH0_JWT_AUDIENCE,
+      });
+      await axiosDefaultInstance.put(`/movies/${movie.movieId}`, putBody, {
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: `Bearer ${token}`,
+        },
+      });
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
   useEffect(() => {
     const handleSetUpFaceDetection = async () => {
       await handleLoadModels();
@@ -102,7 +143,7 @@ const FaceRecognition: FC<FaceRecognitionProps> = ({ movieDetailRef }) => {
     handleSetUpFaceDetection();
 
     return () => {
-      mediaStream.getTracks().forEach((track) => track.stop());
+      if (mediaStream) mediaStream.getTracks().forEach((track) => track.stop());
     };
   }, []);
